Create a branch to backport r1667233 and r1667301 to 1.9.x.
Without r1667301, the r1667233 fix causes an svnserve protocol violation
when the sanity check in get_location_segments() returns an error.  The
backport branch is required to avoid a trunk-specific behavior change in
terms of the error code, i.e., to still use the SVN_ERR_INCORRECT_PARAMS
instead of the SVN_ERR_FS_NO_SUCH_REVISION, as it is now in trunk.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.9.x-r1667233@1673207 13f79535-47bb-0310-9956-ffa450edef68